chenged controllers, readme, added id validation
