+++ controllers/contacts/addContact.js
@@ -1,7 +1,6 @@
-const contacts = require("../../models/contacts");
-
+const { Contact } = require("../../models/contactModel");
 const addContact = async (req, res) => {
-  const contact = await contacts.addContact(req.body);
+  const contact = await Contact.create(req.body);
   res.status(201).json(contact);
 };
