+++ controllers/contacts/removeContact.js
@@ -1,9 +1,9 @@
-const contacts = require("../../models/contacts");
+const { Contact } = require("../../models/contactModel");
 const { RequestError } = require("../../helpers");
 
 const removeContact = async (req, res) => {
   const { contactId } = req.params;
-  const contact = await contacts.removeContact(contactId);
+  const contact = await Contact.findByIdAndRemove(contactId);
   if (!contact) {
     throw RequestError(404, "Not found");
   }
