+++ controllers/contacts/updateContact.js
@@ -1,9 +1,12 @@
-const contacts = require("../../models/contacts");
+const { Contact } = require("../../models/contactModel");
 const { RequestError } = require("../../helpers");
 
 const updateContact = async (req, res) => {
   const { contactId } = req.params;
-  const contact = await contacts.updateContact(contactId, req.body);
+
+  const contact = await Contact.findByIdAndUpdate(contactId, req.body, {
+    new: true,
+  });
   if (!contact) {
     throw RequestError(404, "Not found");
   }
