--- models/contacts.json
@@ -1,65 +1,0 @@
-[
-  {
-    "_id": "6333760025bf9cbe0d2396d3",
-    "name": "Allen Raymond",
-    "email": "nulla.ante@vestibul.co.uk",
-    "phone": "(992) 914-3792",
-    "favorite": false
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396d5",
-    "name": "Kennedy Lane",
-    "email": "mattis.Cras@nonenimMauris.net",
-    "phone": "(542) 451-7038",
-    "favorite": false
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396d6",
-    "name": "Wylie Pope",
-    "email": "est@utquamvel.net",
-    "phone": "(692) 802-2949",
-    "favorite": true
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396d7",
-    "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
-    "phone": "(501) 472-5218",
-    "favorite": true
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396d8",
-    "name": "Abbot Franks",
-    "email": "scelerisque@magnis.org",
-    "phone": "(186) 568-3720",
-    "favorite": true
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396d9",
-    "name": "Reuben Henry",
-    "email": "pharetra.ut@dictum.co.uk",
-    "phone": "(715) 598-5792",
-    "favorite": true
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396da",
-    "name": "Simon Morton",
-    "email": "dui.Fusce.diam@Donec.com",
-    "phone": "(233) 738-2360",
-    "favorite": true
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396db",
-    "name": "Thomas Lucas",
-    "email": "nec@Nulla.com",
-    "phone": "(704) 398-7993",
-    "favorite": false
-  },
-  {
-    "_id": "6333760025bf9cbe0d2396dc",
-    "name": "Alec Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688",
-    "favorite": true
-  }
-]
